+++ src/priority-queue.js
@@ -1,0 +1,4 @@
+/**
+ * @see {PriorityQueue}
+ */
+module.exports = require('./lib/PriorityQueue')